fix(OUT-705) : made listComponent for ListBoxComponent prop of mui autocomplete resuable
