+++ src/components/inputs/Selector.tsx
@@ -2,18 +2,18 @@ import { Avatar, Box, Button, Popper, Stack, Typography } from '@mui/material'
 import { StyledAutocomplete } from '@/components/inputs/Autocomplete'
 import { statusIcons } from '@/utils/iconMatcher'
 import { useFocusableInput } from '@/hooks/useFocusableInput'
-import { HTMLAttributes, ReactNode, Ref, forwardRef, useEffect, useState } from 'react'
+import { HTMLAttributes, ReactNode, useEffect, useState } from 'react'
 import { StyledTextField } from './TextField'
 import { WorkflowStateResponse } from '@/types/dto/workflowStates.dto'
-import { IAssigneeCombined, IExtraOption, ITemplate, UserType, UserTypesName } from '@/types/interfaces'
+import { IAssigneeCombined, IExtraOption, ITemplate, UserTypesName } from '@/types/interfaces'
 import { TruncateMaxNumber } from '@/types/constants'
 
 import { truncateText } from '@/utils/truncateText'
-import { CopilotAvatar } from '../atoms/CopilotAvatar'
+import { CopilotAvatar } from '@/components/atoms/CopilotAvatar'
 import { getAssigneeName } from '@/utils/assignee'
-import { StyledHelperText } from '../error/FormHelperText'
+import { StyledHelperText } from '@/components/error/FormHelperText'
 import React from 'react'
-import Scrollbars, { ScrollbarProps } from 'react-custom-scrollbars'
+import { ListComponent } from '@/components/inputs/ListComponent'
 
 export enum SelectorType {
   ASSIGNEE_SELECTOR = 'assigneeSelector',
@@ -422,37 +422,3 @@ const SelectorButton = ({
     </Button>
   )
 }
-
-interface ListComponentProps extends Omit<ScrollbarProps, 'ref'> {
-  children: React.ReactNode
-}
-
-const ListComponentInternal = forwardRef<Scrollbars, ListComponentProps>((props, ref) => {
-  const { children, ...comProps } = props
-
-  return (
-    <Scrollbars
-      {...comProps}
-      renderThumbVertical={(scrollbarProps) => (
-        <div
-          {...scrollbarProps}
-          className="thumb"
-          style={{
-            borderRadius: '6px',
-            backgroundColor: 'rgba(0, 0, 0, 0.5)',
-          }}
-        />
-      )}
-      autoHeight
-      autoHide
-      autoHideTimeout={500}
-      autoHideDuration={500}
-      hideTracksWhenNotNeeded
-    >
-      {children}
-    </Scrollbars>
-  )
-})
-ListComponentInternal.displayName = 'ListComponentInternal'
-
-const ListComponent = (props: JSX.IntrinsicElements['div']) => <ListComponentInternal {...(props as ListComponentProps)} />
